Revert "research optimal lambda - 2 - auto lambda for bm25"
This reverts commit 73dee143390ac942afde4093ed9b470b60f146f7.

+++ src/main/java/retrieval/KNNRelModel.java
@@ -54,9 +54,9 @@ public class KNNRelModel extends SupervisedRLM {
     public Map<String, List<MsMarcoQuery>> getKnnQueryMap() { return knnQueryMap; }
 
     public float getAvgTopKSimScoreForQ(String qid, int k) { // k for first-k queries to estimate lambda
-        double avgSim = knnQueryMap.get(qid).stream().limit(k).mapToDouble(w -> w.getRefSim()).sum()/k;
+        double avgSim = knnQueryMap.get(qid).stream().limit(k).mapToDouble(w -> double(w.getRefSim())).sum()/k;
 
-        return (float)avgSim;
+        return float(avgSim);
     }
 
     private void constructQueriesAndQrels(String queryFile) throws Exception {
